chat_proj/requirements.txt
Django==5.0.1
django-rest-framework==0.1.0
drf-spectacular==0.27.1
flake8==7.0.0
